+++ Examples/HelloSwiftUsd_CLI_SPM/Sources/main.swift
@@ -19,7 +19,7 @@
 //===----------------------------------------------------------------------===//
 import OpenUSD
 
-public typealias pxr = pxrInternal_v0_26_5__pxrReserved__
+public typealias pxr = pxrInternal_v0_26_8__pxrReserved__
 
 
 func makeHelloWorldString() -> String {